Static Phantom

BDTools.StaticPhantom — Type

A static phantom model

BDTools.staticphantom — Function

staticphantom(ph::Array{Float64, 4}, sliceinfo::Matrix{Int}; staticslices=1:200)

Construct a static phantom object from phantom time series and slice motion data.

BDTools.groundtruth — Function

groundtruth(ph::StaticPhantom, data::AbstractArray, angles::Vector;
            startmotion=1, threshold=Inf, verbose=false)

Construct a prediction of a phantom motion data given a rotation information, and return a tensor of original and predicted values for the masked volume of the phantom ph, slice indices, and mask index map for translation to the original phantom coordinate space.

Ellipse Fitting

BDTools.fitellipse — Function

fitellipse(xy::AbstractMatrix)

Perform an ellipse fitting from a collection of 2D coordinates xy.

Implemented from "Direct least squares fitting of ellipses", Fitzgibbon, 1996.

fitellipse(imgs::AbstractMatrix, segs::SegmentedImage, edge::Matrix{Int};
             verbose=true, keepinitialonerror=true, secondfit=true) -> Vector

Perform fitting of an ellipse given an image tensor imgs, mask for fitting coordinates, and edge elements of an ellipse shape. Returns a vector with fitted ellipse parameters.

If secondfit is true, then an additional fitting is performed after the result of first fitting is corrected.

BDTools.fitellipsedirect — Function

fitellipsedirect(xy::AbstractMatrix)

Perform an ellipse fitting from a collection of 2D coordinates xy.

Implemented from "Numerically stable direct least squares fitting of ellipses", Halir, 1998.

BDTools.canonical — Function

canonical(p::Vector)

Convert ellipse parameters p (6-element vector) in general form into canonical: center coordinates, axes & rotation angle in radians and

Utils

BDTools.getellipse — Function

getellipse(imgs::AbstractArray; verbose=true, secondfit=true) -> Vector

Returns a vector with fitted ellipse parameters from a 2D image matrix.

getellipse(ph::StaticPhantom, z::Int)

Return ellipse properties for a slice z of a static phantom ph.

BDTools.ellipserot — Function

ellipserot(α, θ, a, b)

Ellipsoidal rotation matrix:

α: rotation angle

θ: ellipse initial rotation angle

a: ellipse long axis

b: ellipse short axis

BDTools.rotate — Function

rotate(x, y, z, a, b, c, u, v, w, θ)

Rotate point (x,y,z) around the line that passes through the point (a,b,c) in the direction of <u,v,w> on the angle θ.

BDTools.rotatevoxel — Function

rotatevoxel(o::Vector, p::Vector) -> f(x,y,z,θ)

Return a function that perform rotation of the voxel with coordinates (x,y,z) on the angle θ within XY-plane around the line started in an origin point o, and passing through a point p.

BDTools.simulated_coordinates — Function

simulated_coordinates(sz::Tuple, a::Vector, b::Vector, θ::Float64)

Generate simulated coordinate set of size sz by rotating voxels within xy-plain on angle θ around the line passing through points a and b.

BDTools.phantominterp — Function

phantominterp(imgs::AbstractArray; itrptype = BSpline(Quadratic()))

Construct interpolation function from an image 3D tensor.

BDTools.fitline — Function

fitline(xy::AbstractMatrix) -> NamedTuple

Return line paramaters (mean, direction, slope) after fiting coordinates stored in column-major order input xy.

BDTools.getangles — Function

getangles(file::String; initpos=20, col=11)

Read phantom rotation data from a file and extract rotation angle in radians. Return vector of angles, and an index in it of a first valid rotation.