Static Phantom

BDTools.StaticPhantom — Type

A static phantom model

BDTools.staticphantom — Function

staticphantom(ph::Array{Float64, 4}, sliceinfo::Matrix{Int}; staticslices=1:200)

Construct a static phantom object from phantom time series ph and slice motion data sliceinfo. The parameter staticslices allows to select a number slices are used for construction of a model.

BDTools.findinitialrotation — Method

findinitialrotation(sph::StaticPhantom, z::Real; intensity=0.01)

Find an ellipse initial rotation angle that minimizes distortions from elliptical rotation. An intensity parameter specifies an intensity percentage for selection of points as parameters for optimization task that determines phantom initial rotation.

BDTools.mask — Method

mask(ph::StaticPhantom, z::Int; threshold=1.0)

Get coordinates of an internal segment of the phantom's ph slice z. The threshold parameter allow to specify a percentage of masked coordinates.

BDTools.GroundTruth — Type

Phantom Simulated Ground Truth

A ground truth representation of a phantom motion data given a rotation information. It contains a tensor of original and predicted values for the masked volume of an original phantom data, a slice index map, and a mask index map for translation to the original phantom coordinate space.

BDTools.groundtruth — Function

groundtruth(ph::StaticPhantom, data::AbstractArray, angles::Vector;
            startmotion=1, threshold=Inf, verbose=false)

Construct a prediction of a phantom motion data given a rotation information, and return a tensor of original and predicted values for the masked volume of the phantom ph, slice indices, and mask index map for translation to the original phantom coordinate space.

Base.getindex — Method

getindex(gt::GroundTruth, x::Int, y::Int)

Return an index of a masked phantom volume.

Base.getindex — Method

getindex(gt::GroundTruth, x::Int, y::Int, z::Int, original::Bool=false)

Return a phantom predicted ground truth data in the coordinate (x,y,z). If original is set to true, original phantom data is returned.

I/O

BDTools.serialize — Function

serialize(filepath::String, gt::GroundTruth)

Serialize a phantom ground truth data object gt into HDF5-formatted file.

Example:

julia> BDTools.serialize("gt.h5", gt)

BDTools.deserialize — Function

deserialize(filepath::String, ::Type{GroundTruth})

Deserialize a phantom ground truth data from a HDF5-formatted file, and construct GroundTruth object.

Example:

julia> gt = BDTools.deserialize("gt.h5", GroundTruth)
GroundTruth: (598, 359, 9, 2)

Ellipse Fitting

BDTools.fitellipse — Function

fitellipse(xy::AbstractMatrix)

Perform an ellipse fitting from a collection of 2D coordinates xy.

Implemented from "Direct least squares fitting of ellipses", Fitzgibbon, 1996.

fitellipse(imgs::AbstractMatrix, segs::SegmentedImage, edge::Matrix{Int};
             verbose=true, keepinitialonerror=true, secondfit=true) -> Vector

Perform fitting of an ellipse given an image tensor imgs, mask for fitting coordinates, and edge elements of an ellipse shape. Returns a vector with fitted ellipse parameters.

If secondfit is true, then an additional fitting is performed after the result of first fitting is corrected.

BDTools.fitellipsedirect — Function

fitellipsedirect(xy::AbstractMatrix)

Perform an ellipse fitting from a collection of 2D coordinates xy.

Implemented from "Numerically stable direct least squares fitting of ellipses", Halir, 1998.

BDTools.canonical — Function

canonical(p::Vector)

Convert ellipse parameters p (6-element vector) in general form into canonical: center coordinates, axes & rotation angle in radians and

Utils

BDTools.getellipse — Function

getellipse(imgs::AbstractArray; verbose=true, secondfit=true) -> Vector

Returns a vector with fitted ellipse parameters from a 2D image matrix.

getellipse(ph::StaticPhantom, z::Int)

Return ellipse properties for a slice z of a static phantom ph.

BDTools.ellipserot — Function

ellipserot(α, θ, a, b)

Ellipsoidal rotation matrix:

α: rotation angle

θ: ellipse initial rotation angle

a: ellipse long axis

b: ellipse short axis

BDTools.rotate — Function

rotate(x, y, z, a, b, c, u, v, w, θ)

Rotate point (x,y,z) around the line that passes through the point (a,b,c) in the direction of <u,v,w> on the angle θ.

BDTools.rotatevoxel — Function

rotatevoxel(o::Vector, p::Vector) -> f(x,y,z,θ)

Return a function that perform rotation of the voxel with coordinates (x,y,z) on the angle θ within XY-plane around the line started in an origin point o, and passing through a point p.

BDTools.simulated_coordinates — Function

simulated_coordinates(sz::Tuple, a::Vector, b::Vector, θ::Float64)

Generate simulated coordinate set of size sz by rotating voxels within xy-plain on angle θ around the line passing through points a and b.

BDTools.phantominterp — Function

phantominterp(imgs::AbstractArray; itrptype = BSpline(Quadratic()))

Construct interpolation function from an image 3D tensor.

BDTools.fitline — Function

fitline(xy::AbstractMatrix) -> NamedTuple

Return line paramaters (mean, direction, slope) after fiting coordinates stored in column-major order input xy.

BDTools.getangles — Function

getangles(file::String; initpos=20, col=11)

Read phantom rotation data from a file and extract rotation angle in radians. Return vector of angles, and an index in it of a first valid rotation.

BDTools.maskindex — Function

maskindex(mask::BitMatrix)

Construct a coordinate index for phantom slice mask.

QualityMeasures

BDTools.st_snr — Function

st_snr(pred_ts, orig_ts)

Calculates the standardized signal-to-noise ratio: Power(signal)/Power(noise)

BDTools.mul_noise — Function

mulnoise(predts, orig_ts)

Calculates the amplitudes of white and multiplicative noise using HMC sampling

Denoiser

BDTools.Denoiser.load_ground_truth — Function

load_ground_truth(fname::String)

Returns a GroundTruth object, a standardized ground truth data, and standardization parameters, i.e. μ and σ.

BDTools.Denoiser.load_phantom — Function

load_phantom(gt::GroundTruth, phantomfile::String)

Return standardized phantom time series data, and standardization parameters, i.e. μ and σ.

BDTools.Denoiser.DenoiseNet — Type

A wrapper type for denoiser model. It contains deep neural network parameters, a loss function, and training parameters.

BDTools.Denoiser.TrainParameters — Type

Training parameters for a denoising network.

adj_rate: a learning rate adjustment factor (default value is 0.33)

batch_size: a training data batch size (default value is 8)

epochs: a number of training epochs (default value is 250)

eta: a learning rate (default value is 3e-4)

kernel_size: a convolutional layer kernel size (default value is 9)

layers: a number of denoiser network convolutional layers (default value is 6)

save_path: a path to a location where a NN model is saved

save_best: if true then the denoiser model is saved on each successful loss improvement

seed: a PRNG seed (default value is 180181)

test_split: a fraction of a phantom data dedicated to a testing set (default value is 0.33)

BDTools.Denoiser.train! — Function

train!(denoiser::Denoiser, sim::AbstractArray{T,3}, ori::AbstractArray{T,3})

Train a denoiser network given a ground truth data, sim, and an original phantom data, ori. The data has to be standardized.

BDTools.Denoiser.denoise — Function

denoise(model::DenoiseNet, data)

Denoise data using a denoseing CNN model. The data input should be standardized.